requirements.txt
Django>=5.2.7
djangorestframework
djoser
djangorestframework-simplejwt
python-dotenv
Pillow
whitenoise
